Cleaned up helpers and interfaces
Moved interfaces to interfaces.ts

Added AliasPath interface

Update tsconfig.json

Added config object

Refactored helpers

cleaned up code

Refactored alias setup

style: ran formatting script

Reexport ReplaceTscAliasPathsOptions interface

formatting

cleaned up trie
