+++ src/index.ts
@@ -11,11 +11,13 @@ import {
   resolve
 } from 'path';
 import {
-  existsResolvedAlias,
-  getAbsoluteAliasPath,
-  getProjectDirPathInOutDir,
-  loadConfig
+  findBasePathOfAlias,
+  loadConfig,
+  relativeOutPathToConfigDir,
+  replaceBaseUrlImport,
+  replaceImportStatement
 } from './helpers';
+import { ReplaceTscAliasPathsOptions, Alias, IConfig } from './interfaces';
 import {
   newStringRegex,
   Output,
@@ -24,24 +26,8 @@ import {
   TrieNode
 } from './utils';
 
-export interface ReplaceTscAliasPathsOptions {
-  configFile?: string;
-  outDir?: string;
-  watch?: boolean;
-  silent?: boolean;
-  resolveFullPaths?: boolean;
-}
-
-interface Alias {
-  shouldPrefixMatchWildly: boolean;
-  prefix: string;
-  basePath: string;
-  path: string;
-  paths: string[];
-  isExtra: boolean;
-}
-
-type Assertion = (claim: any, message: string) => asserts claim;
+// export interfaces for api use.
+export { ReplaceTscAliasPathsOptions };
 
 export async function replaceTscAliasPaths(
   options: ReplaceTscAliasPathsOptions = {
@@ -58,226 +44,66 @@ export async function replaceTscAliasPaths(
     ? resolve(process.cwd(), options.configFile)
     : options.configFile;
 
-  const assert: Assertion = (claim, message) =>
-    claim || output.error(message, true);
-
-  assert(existsSync(configFile), `Invalid file path => ${configFile}`);
+  output.assert(existsSync(configFile), `Invalid file path => ${configFile}`);
 
   let { baseUrl = './', outDir, paths } = loadConfig(configFile);
   if (options.outDir) outDir = options.outDir;
 
-  assert(outDir, 'compilerOptions.outDir is not set');
+  output.assert(outDir, 'compilerOptions.outDir is not set');
 
   const configDir: string = normalizePath(dirname(configFile));
 
-  const outPath = normalizePath(normalize(configDir + '/' + outDir));
-
-  const confDirParentFolderName: string = basename(configDir);
-
-  let hasExtraModule = false;
-  let configDirInOutPath: string = null;
-  let relConfDirPathInOutPath: string;
+  const config: IConfig = {
+    configFile: configFile,
+    baseUrl: baseUrl,
+    outDir: outDir,
+    configDir: configDir,
+    outPath: normalizePath(normalize(configDir + '/' + outDir)),
+    confDirParentFolderName: basename(configDir),
+    hasExtraModule: false,
+    configDirInOutPath: null,
+    relConfDirPathInOutPath: null
+  };
 
   const AliasTrie = new TrieNode<Alias>();
 
   Object.keys(paths)
     .map((alias) => {
-      const _paths = paths[alias].map((path) => {
-        path = path.replace(/\*$/, '').replace(/\.([mc])?ts(x)?$/, '.$1js$2');
-        if (isAbsolute(path)) {
-          path = relative(configDir, path);
-        }
-        return path;
-      });
-
-      const path = _paths[0];
-
-      if (normalize(path).includes('..')) {
-        if (!configDirInOutPath) {
-          configDirInOutPath = getProjectDirPathInOutDir(
-            outPath,
-            confDirParentFolderName
-          );
-
-          // Find relative path access of configDir in outPath
-          if (configDirInOutPath) {
-            hasExtraModule = true;
-            const stepsbackPath = relative(configDirInOutPath, outPath);
-            const splitStepBackPath = normalizePath(stepsbackPath).split('/');
-            const nbOfStepBack = splitStepBackPath.length;
-            const splitConfDirInOutPath = configDirInOutPath.split('/');
-
-            let i = 1;
-            const splitRelPath: string[] = [];
-            while (i <= nbOfStepBack) {
-              splitRelPath.unshift(
-                splitConfDirInOutPath[splitConfDirInOutPath.length - i]
-              );
-              i++;
-            }
-            relConfDirPathInOutPath = splitRelPath.join('/');
-          }
-        }
-      }
-
       return {
         shouldPrefixMatchWildly: alias.endsWith('*'),
         prefix: alias.replace(/\*$/, ''),
-        basePath: null,
-        path,
-        paths: _paths,
-        isExtra: null
+        // Normalize paths.
+        paths: paths[alias].map((path) => {
+          path = path.replace(/\*$/, '').replace(/\.([mc])?ts(x)?$/, '.$1js$2');
+          if (isAbsolute(path)) {
+            path = relative(config.configDir, path);
+          }
+
+          if (normalize(path).includes('..') && !config.configDirInOutPath) {
+            relativeOutPathToConfigDir(config);
+          }
+
+          return path;
+        })
       };
     })
-    .filter(({ prefix }) => prefix)
-    /*********** Find basepath of aliases *****************/
     .forEach((alias) => {
-      if (normalize(alias.path).includes('..')) {
-        const tempBasePath = normalizePath(
-          normalize(
-            `${configDir}/${outDir}/${
-              hasExtraModule && relConfDirPathInOutPath
-                ? relConfDirPathInOutPath
-                : ''
-            }/${baseUrl}`
-          )
-        );
-
-        const absoluteBasePath = normalizePath(
-          normalize(`${tempBasePath}/${alias.path}`)
-        );
-        if (existsResolvedAlias(absoluteBasePath)) {
-          alias.isExtra = false;
-          alias.basePath = tempBasePath;
-        } else {
-          alias.isExtra = true;
-          alias.basePath = absoluteBasePath;
-        }
-      } else if (hasExtraModule) {
-        alias.isExtra = false;
-        alias.basePath = normalizePath(
-          normalize(
-            `${configDir}/${outDir}/${relConfDirPathInOutPath}/${baseUrl}`
-          )
-        );
-      } else {
-        alias.basePath = normalizePath(normalize(`${configDir}/${outDir}`));
-        alias.isExtra = false;
+      if (alias.prefix) {
+        // Add all aliases to AliasTrie.
+        AliasTrie.add(alias.prefix, {
+          ...alias,
+          // Find basepath of aliases.
+          paths: alias.paths.map(findBasePathOfAlias(config))
+        });
       }
-
-      // Add all aliases to AliasTrie.
-      AliasTrie.add(alias.prefix, alias);
     });
 
-  const replaceImportStatement = ({
-    orig,
-    file,
-    alias
-  }: {
-    orig: string;
-    file: string;
-    alias: Alias;
-  }): string => {
-    const requiredModule = orig.match(newStringRegex())?.groups?.path;
-    assert(
-      typeof requiredModule == 'string',
-      `Unexpected import statement pattern ${orig}`
-    );
-    const isAlias = alias.shouldPrefixMatchWildly
-      ? // if the alias is like alias*
-        // beware that typescript expects requiredModule be more than just alias
-        requiredModule.startsWith(alias.prefix) &&
-        requiredModule !== alias.prefix
-      : // need to be a bit more careful if the alias doesn't ended with a *
-        // in this case the statement must be like either
-        // require('alias') or require('alias/path');
-        // but not require('aliaspath');
-        requiredModule === alias.prefix ||
-        requiredModule.startsWith(alias.prefix + '/');
-
-    if (isAlias) {
-      for (let i = 0; i < alias.paths.length; i++) {
-        const absoluteAliasPath = getAbsoluteAliasPath(
-          alias.basePath,
-          alias.paths[i]
-        );
-
-        // Check if path is valid.
-        if (
-          !existsResolvedAlias(
-            alias.prefix.length == requiredModule.length
-              ? normalizePath(absoluteAliasPath)
-              : normalizePath(
-                  `${absoluteAliasPath}/${requiredModule.replace(
-                    new RegExp(`^${alias.prefix}`),
-                    ''
-                  )}`
-                )
-          )
-        ) {
-          continue;
-        }
-
-        let relativeAliasPath: string = normalizePath(
-          relative(dirname(file), absoluteAliasPath)
-        );
-
-        if (!relativeAliasPath.startsWith('.')) {
-          relativeAliasPath = './' + relativeAliasPath;
-        }
-
-        const index = orig.indexOf(alias.prefix);
-        const newImportScript =
-          orig.substring(0, index) +
-          relativeAliasPath +
-          '/' +
-          orig.substring(index + alias.prefix.length);
-
-        const modulePath = newImportScript.match(newStringRegex()).groups.path;
-
-        return newImportScript.replace(modulePath, normalizePath(modulePath));
-      }
-    }
-    return orig;
-  };
-
-  const replaceBaseUrlImport = ({
-    orig,
-    file
-  }: {
-    orig: string;
-    file: string;
-  }): string => {
-    const requiredModule = orig.match(newStringRegex())?.groups?.path;
-    assert(
-      typeof requiredModule == 'string',
-      `Unexpected import statement pattern ${orig}`
-    );
-
-    // Check if import is already resolved.
-    if (requiredModule.startsWith('.')) {
-      return orig;
-    }
-
-    // If there are files matching the target, resolve the path.
-    if (existsResolvedAlias(`${outPath}/${requiredModule}`)) {
-      let relativePath: string = normalizePath(
-        relative(dirname(file), getAbsoluteAliasPath(outPath, ''))
-      );
-      if (!relativePath.startsWith('.')) {
-        relativePath = './' + relativePath;
-      }
-
-      const index = orig.indexOf(requiredModule);
-      const newImportScript =
-        orig.substring(0, index) + relativePath + '/' + orig.substring(index);
-
-      const modulePath = newImportScript.match(newStringRegex()).groups.path;
-      return newImportScript.replace(modulePath, normalizePath(modulePath));
-    }
-    return orig;
-  };
-
+  /**
+   * replaceAlias replaces aliases in file.
+   * @param file file to replace aliases in.
+   * @param resolveFullPath if tsc-alias should resolve the full path
+   * @returns if something has been replaced.
+   */
   const replaceAlias = async (
     file: string,
     resolveFullPath?: boolean
@@ -292,19 +118,26 @@ export async function replaceTscAliasPaths(
       );
       // If an alias is found replace it or return the original.
       return alias
-        ? replaceImportStatement({
-            orig,
-            file,
-            alias
-          })
+        ? replaceImportStatement(
+            {
+              orig,
+              file,
+              alias
+            },
+            output.assert
+          )
         : orig;
     });
 
     tempCode = replaceSourceImportPaths(tempCode, file, (orig) =>
-      replaceBaseUrlImport({
-        orig,
-        file
-      })
+      replaceBaseUrlImport(
+        {
+          orig,
+          file
+        },
+        output.assert,
+        config
+      )
     );
 
     // Fully resolve all import paths (not just aliased ones)
@@ -322,8 +155,8 @@ export async function replaceTscAliasPaths(
 
   // Finding files and changing alias paths
   const globPattern = [
-    `${outPath}/**/*.{mjs,cjs,js,jsx,d.{mts,cts,ts,tsx}}`,
-    `!${outPath}/**/node_modules`
+    `${config.outPath}/**/*.{mjs,cjs,js,jsx,d.{mts,cts,ts,tsx}}`,
+    `!${config.outPath}/**/node_modules`
   ];
   const files = sync(globPattern, {
     dot: true,
@@ -346,7 +179,7 @@ export async function replaceTscAliasPaths(
   if (options.watch) {
     output.info('[Watching for file changes...]');
     const filesWatcher = watch(globPattern);
-    const tsconfigWatcher = watch(configFile);
+    const tsconfigWatcher = watch(config.configFile);
     const onFileChange = async (file: string) =>
       await replaceAlias(file, options?.resolveFullPaths);
     filesWatcher.on('add', onFileChange);
