+++ src/utils/output.ts
@@ -8,12 +8,16 @@ export class Output {
 
   error(message: string, exitProcess = false) {
     if (!this.silent)
-      //             [BgRed_] Error: [Reset] [FgRed_]${message}[Reset][LF]
+      //           [BgRed_] Error: [Reset] [FgRed_]${message}[Reset][LF]
       console.log(`\x1b[41m Error: \x1b[0m \x1b[31m${message}\x1b[0m\n`);
     if (exitProcess) process.exit(1);
   }
 
   clear() {
     console.clear();
+  }
+
+  assert(claim: any, message: string) {
+    claim || this.error(message, true);
   }
 }
