+++ src/utils/trie.ts
@@ -39,15 +39,20 @@ export class TrieNode<T> {
     if (name.length <= 0) return null;
 
     const node = this.children.get(name[0]);
-    if (node) {
-      if (name.length == 1) {
-        return node.data;
-      } else {
-        const result = node.search(name.substring(1));
-        return result ? result : node.data;
-      }
-    } else {
-      return this.data;
-    }
+    return node
+      ? name.length == 1
+        ? node.data
+        : this.compare(node.search(name.substring(1)), node.data)
+      : this.data;
+  }
+
+  /**
+   * compares a and b.
+   * @param a first element to compare.
+   * @param b second element to compare.
+   * @returns first element if true otherwise the second.
+   */
+  private compare(a: T | null, b: T | null): T | null {
+    return a ? a : b;
   }
 }
